third_party/WebKit/LayoutTests/dom/document_fragment/sibling-rules-under-document-fragment-expected.txt

Tests :first-child rule under DocumentFragment.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fragmentSelected is a1
PASS elementSelected is a2
PASS successfullyParsed is true

TEST COMPLETE
